added token routes

+++ routes/tasks.js
@@ -4,13 +4,10 @@ module.exports = app => {
 
     app.route("/tasks")
         // Middleware for pre-execution of routes
-        // .all((req, res, next) => {
-        //     delete req.body.id;  // ensures the exclusion of the id attr within the request's body
-        //     next();
-        // })
+        .all(app.auth.authenticate())
         
         .get((req, res) => {
-            Tasks.findAll({})
+            Tasks.findAll({ where: {user_id: req.user.id}})
                 .then(result => res.json(result))
                 .catch(error => {
                     res.status(412).json({ msg: error.message});
@@ -18,6 +15,7 @@ module.exports = app => {
         })
 
         .post((req, res) => {
+            req.body.user_id = req.user.id // req.user.id is data sent by done callback in auth.js
             Tasks.create(req.body)
                 .then(result => res.json(result))
                 .catch(error => {
@@ -28,13 +26,10 @@ module.exports = app => {
     
 // single task
     app.route("/tasks/:id")
-        // .all((req, res, next) => {
-        //     delete req.body.id;
-        //     next();
-        // })
-
+        .all(app.auth.authenticate())
+        
         .get((req, res) => {
-            Tasks.findOne({where: req.params})
+            Tasks.findOne({where:{ id: req.params.id, user_id: req.user.id }})
                 .then(result => {
                     if (result) {
                         res.json(result)
@@ -49,7 +44,7 @@ module.exports = app => {
         })
 
         .put((req, res) => {
-            Tasks.update(req.body, {where: req.params})
+            Tasks.update(req.body, {where: {id: req.params.id, user_id: req.user.id}})
             .then(result => res.sendStatus(204))  //204 means successful but no response
             .catch(error => {
                 res.status(412).json({ msg: error.message});
@@ -57,7 +52,7 @@ module.exports = app => {
         })
 
         .delete((req, res) => {
-            Tasks.destroy({ where: req.params})
+            Tasks.destroy({ where: {id: req.params.id, user_id: req.user.id}})
             .then(result => res.sendStatus(204))
             .catch(error => {
                 res.status(412).json({ msg: error.message})
